+++ lbrynet/reflector/__init__.py
@@ -1,4 +1,4 @@
 from lbrynet.reflector.server.server import ReflectorServerFactory as ServerFactory
 from lbrynet.reflector.client.client import EncryptedFileReflectorClientFactory as ClientFactory
 from lbrynet.reflector.client.client import BlobReflectorClientFactory as BlobClientFactory
-from lbrynet.reflector.util import ReflectorAvailabilityHelper
+from lbrynet.reflector import reupload
